iniciando o refactor da componentização do sistema
